requirements.txt
flask>=3.0.0
pyfiglet>=1.0.2
Pillow>=10.0.0
huggingface-hub>=0.20.0
textual>=0.47.0
rich>=13.7.0
